clean up and disabling iv rotation for now
